requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
python-dotenv
httpx
